Refactor: Prune goroutines and condense broadcast functions
